+++ src/components/fantasy-leagues/my-team/MyTeamEditView.tsx
@@ -1,7 +1,6 @@
 import { Fragment } from 'react/jsx-runtime';
-import { FANTASY_TEAM_POSITIONS, MAX_TEAM_BUDGET } from '../../../types/constants';
+import { MAX_TEAM_BUDGET } from '../../../types/constants';
 import { IFantasyTeamAthlete } from '../../../types/fantasyTeamAthlete';
-import { PlayerGameCard } from '../../player/PlayerGameCard';
 import { useEffect, useMemo, useState } from 'react';
 import { IFantasyLeagueRound, IFantasyLeagueTeam } from '../../../types/fantasyLeague';
 import { fantasyTeamService } from '../../../services/fantasyTeamService';
@@ -9,17 +8,13 @@ import { seasonService } from '../../../services/seasonsService';
 import { IProAthlete } from '../../../types/athletes';
 import { Position } from '../../../types/position';
 import { IGamesLeagueConfig } from '../../../types/leagueConfig';
-import { Loader, Check, X } from 'lucide-react';
+import { Loader, Check } from 'lucide-react';
 import PlayerProfileModal from '../../player/PlayerProfileModal';
 import PrimaryButton from '../../shared/buttons/PrimaryButton';
 import PlayerSelectionModal from '../../team-creation/PlayerSelectionModal';
-import { isLeagueRoundLocked } from '../../../utils/leaguesUtils';
 import { useFantasyLeagueTeam } from './FantasyLeagueTeamProvider';
 import { IFantasyLeagueTeamSlot } from '../../../types/fantasyLeagueTeam';
-import { useFantasyLeague } from '../../fantasy-league/useFantasyLeague';
-import { useFantasyLeagueGroup } from '../../../hooks/leagues/useFantasyLeagueGroup';
-import SecondaryText from '../../shared/SecondaryText';
-import { twMerge } from 'tailwind-merge';
+import { EditableTeamSlotItem } from './EditableTeamSlotItem';
 
 type Props = {
   leagueRound?: IFantasyLeagueRound;
@@ -29,20 +24,13 @@ type Props = {
   onEditChange?: (isEditing: boolean) => void;
 };
 /** Renders My Team Edit Grid */
-export default function MyTeamEditView({
-  team,
-  leagueConfig,
-  leagueRound,
-  onEditChange,
-  onTeamUpdated,
-}: Props) {
+export default function MyTeamEditView({ team, leagueConfig,leagueRound, onTeamUpdated }: Props) {
 
   const {
-    slots, removePlayerAtSlot,
-    setPlayerAtSlot, teamCaptain,
-    totalSpent, setTeamCaptainAtSlot,
-    resetToOriginalTeam, originalSlots,
-    originalCaptain, setOldPlayerAtSlot
+    slots,
+    setPlayerAtSlot, teamCaptain, totalSpent,
+    resetToOriginalTeam, originalSlots, originalCaptain, 
+    changesDetected, isTeamFull
   } = useFantasyLeagueTeam();
 
   const [playerModalPlayer, setPlayerModalPlayer] = useState<IFantasyTeamAthlete>();
@@ -73,20 +61,7 @@ export default function MyTeamEditView({
   const budgetRemaining = (leagueConfig?.team_budget || MAX_TEAM_BUDGET) - totalSpent;
 
   const isEditing = useMemo(() => {
-    // Captain changed?
-    if (teamCaptain?.tracking_id !== originalCaptain?.tracking_id) return true;
-
-    // Any slot player changed?
-    for (let i = 1; i <= slots.length; i++) {
-      const slot = slots[i];
-      const ogSlot = originalSlots[i];
-
-      const orig = ogSlot?.athlete?.athlete_id || undefined;
-      const curr = slot?.athlete?.athlete_id || undefined;
-      if (orig !== curr) return true;
-    }
-    return false;
-
+    return changesDetected;
   }, [teamCaptain, originalSlots, originalCaptain, slots]);
 
   // Cancel: revert to original team state
@@ -169,13 +144,13 @@ export default function MyTeamEditView({
     }
   };
 
-  const handleSetCaptain = (slotNumber: number) => {
-    setTeamCaptainAtSlot(slotNumber);
-  }
+  const handleCompleteSwapPlayer = (newAthlete: IProAthlete) => {
+    
+    console.log("About to swap player", swapState);
+    
+    if (!swapState || swapState.slot === null || !swapPlayer?.slot === undefined) return;
 
-  const handleSwapPlayer = (newAthlete: IProAthlete) => {
-    if (!swapState || !swapState.slot) return;
-
+    console.log("Yey, so we passed the test bro");
     setPlayerAtSlot(swapState.slot, newAthlete);
     setSwapState({ open: false, slot: null, position: null });
   }
@@ -207,9 +182,19 @@ export default function MyTeamEditView({
       open: true,
       slot: slot.slotNumber,
       position: pos
-    })
+    });
 
+    setSwapPlayer(undefined);
   }
+
+  // const handleClearSlot = (slot: IFantasyLeagueTeamSlot) => {
+  //   const {playerRemoved} = removePlayerAtSlot(slot.slotNumber);
+  //   if (playerRemoved) {
+  //     setSwapState({
+
+  //     })
+  //   }
+  // }
 
   return (
     <Fragment>
@@ -226,7 +211,7 @@ export default function MyTeamEditView({
             </button>
             <PrimaryButton
               className="w-1/2"
-              disabled={isSaving || !leagueRound?.is_open}
+              disabled={isSaving || !leagueRound?.is_open || !isTeamFull}
               onClick={buildPayloadAndSave}
             >
               {isSaving ? 'Saving...' : 'Save Changes'}
@@ -282,7 +267,7 @@ export default function MyTeamEditView({
                 })
             }
 
-            handlePlayerSelect={handleSwapPlayer}
+            handlePlayerSelect={handleCompleteSwapPlayer}
             onClose={handleCancelSwap}
             roundId={parseInt(String(leagueRound?.id || '0'))}
             roundStart={leagueRound?.start_round ?? undefined}
@@ -343,161 +328,3 @@ export default function MyTeamEditView({
     </Fragment >
   );
 }
-
-type SlotProps = {
-  slot: IFantasyLeagueTeamSlot,
-  onPlayerClick?: (player: IFantasyTeamAthlete) => void;
-  disabled?: boolean;
-  onInitiateSwap?: (slot: IFantasyLeagueTeamSlot) => void;
-  onAddPlayerToEmptySlot?: (slot: IFantasyLeagueTeamSlot) => void;
-}
-
-function EditableTeamSlotItem({ slot, onPlayerClick, disabled, onInitiateSwap, onAddPlayerToEmptySlot }: SlotProps) {
-
-  const { currentRound } = useFantasyLeagueGroup();
-  const { setTeamCaptainAtSlot, removePlayerAtSlot } = useFantasyLeagueTeam();
-  const athlete = slot.athlete;
-
-  const isCurrPlayerCaptain = slot.isCaptain;
-
-  const handlePlayerClick = () => {
-    if (onPlayerClick && slot.athlete) {
-      onPlayerClick(slot.athlete);
-    }
-  }
-
-  const handleSetCaptain = () => {
-    setTeamCaptainAtSlot(slot.slotNumber);
-  }
-
-  const handleInitiateSwap = () => {
-    if (onInitiateSwap) {
-      onInitiateSwap(slot);
-    }
-  }
-
-  const handleClearSlot = () => {
-    removePlayerAtSlot(slot.slotNumber);
-  }
-
-  const cannotSelectCaptain = disabled || isCurrPlayerCaptain;
-  const isLocked = currentRound && isLeagueRoundLocked(currentRound);
-
-  return (
-    <Fragment>
-
-      <div key={athlete?.tracking_id} className="flex flex-col w-full min-w-0 p-2">
-        <div className="w-full min-w-0 h-60 flex items-center justify-center bg-transparent">
-          {athlete ? (
-            <div className="w-full h-full flex flex-col items-center justify-center">
-
-              <div className=' flex w-full flex-row items-center justify-between' >
-                <SecondaryText>{slot.position.name}</SecondaryText>
-                <div>
-                  <button
-                    onClick={handleClearSlot}
-                    className='dark:bg-slate-700/60 bg-slate-200 hover:dark:bg-slate-700 w-6 h-6 rounded-md flex flex-col items-center justify-center'
-                  >
-                    <X className='w-4 h-4 text-slate-700 dark:text-white' />
-                  </button>
-                </div>
-              </div>
-
-              <PlayerGameCard
-                key={athlete.tracking_id}
-                player={athlete}
-                className="mx-auto"
-                blockGlow
-                onClick={handlePlayerClick}
-                detailsClassName="pl-6 pr-6 pb-7"
-                priceClassName="top-12 left-6"
-                teamLogoClassName="top-4 right-2"
-              />
-            </div>
-          ) : (
-            <EmptySlotCard
-              slot={slot}
-              onClickSlot={onAddPlayerToEmptySlot}
-            />
-          )}
-        </div>
-
-        {athlete && (
-          <div className="mt-4 flex flex-col gap-2 z-50">
-            <button
-              className={`${isCurrPlayerCaptain
-                ? 'text-xs w-full rounded-lg py-2 bg-amber-100 dark:bg-amber-900/30 text-amber-800 dark:text-amber-300 border border-amber-200 dark:border-amber-700'
-                : `text-xs w-full rounded-lg py-2 bg-blue-50 dark:bg-blue-900/30 text-blue-700 dark:text-blue-300 border border-blue-200 dark:border-blue-700 ${isLocked ? '' : 'hover:bg-blue-100 dark:hover:bg-blue-900/50'}`
-                }`}
-              onClick={handleSetCaptain}
-              disabled={cannotSelectCaptain}
-            >
-              {isCurrPlayerCaptain ? 'Captain' : 'Make Captain'}
-            </button>
-
-            <button
-              className={`text-xs w-full rounded-lg py-2 bg-purple-50 dark:bg-purple-900/30 text-purple-700 dark:text-purple-300 border border-purple-200 dark:border-purple-700 ${isLocked ? '' : 'hover:bg-purple-100 dark:hover:bg-purple-900/50 disabled:opacity-60'}`}
-              onClick={handleInitiateSwap}
-              disabled={disabled}
-            >
-              Swap
-            </button>
-          </div>
-        )}
-
-        {!athlete && (
-          <div className="mt-4 flex flex-col gap-2 z-50">
-            <button
-              className={`text-xs w-full rounded-lg py-2 bg-purple-50 dark:bg-purple-900/30 text-purple-700 dark:text-purple-300 border border-purple-200 dark:border-purple-700 ${isLocked ? '' : 'hover:bg-purple-100 dark:hover:bg-purple-900/50 disabled:opacity-60'}`}
-              onClick={() => {
-                if (onAddPlayerToEmptySlot) {
-                  onAddPlayerToEmptySlot(slot)
-                }
-              }}
-              disabled={disabled}
-            >
-              Add Player
-            </button>
-          </div>
-        )}
-      </div>
-
-    </Fragment>
-  )
-}
-
-type EmptySlotCardProps = {
-  slot: IFantasyLeagueTeamSlot,
-  onClickSlot?: (slot: IFantasyLeagueTeamSlot) => void
-}
-
-function EmptySlotCard({ slot, onClickSlot }: EmptySlotCardProps) {
-
-  const handleClickSlot = () => {
-    if (onClickSlot) {
-      onClickSlot(slot);
-    }
-  }
-
-  return (
-    <div
-      onClick={handleClickSlot}
-      className={twMerge(
-        "flex flex-col cursor-pointer hover:dark:bg-slate-800/40 dark:bg-slate-800/20 rounded-xl border-4 border-slate-500/30 items-center justify-center border-dotted w-full h-full",
-        "bg-white hover:bg-gray-100"
-      )}
-    >
-      <span className="text-3xl">+</span>
-      <span className="mt-2 text-xs text-gray-500 dark:text-gray-400">{slot.position.name}</span>
-    </div>
-  )
-}
-
-// onClick={() => {
-//   const pos = toPosition(positions[index], index);
-//   setSwapState({ open: true, slot: slotNumber, position: pos });
-
-//   if (s.athlete) {
-//     setSwapPlayer(s.athlete);
-//   }
-// }}
